Fix mac warning about size_t and %d.

+++ lib/Transforms/ForEachExpand.cpp
@@ -185,7 +185,8 @@ namespace {
     }
 
     if (ArgVec.size() != 5) {
-      ALOGE("Incorrect number of arguments to function: %d", ArgVec.size());
+      ALOGE("Incorrect number of arguments to function: %d",
+          (int) ArgVec.size());
       return false;
     }
     llvm::Value *Arg_p = ArgVec[0];
